Sorting moved away from VoyagesList Component
